refactor: rename tasks-files provider to context
Renamed `provider` to `context` in tasks-files components for consistency with established naming conventions. Updated all imports and exports to reflect the change.
